python-chunking-service/requirements.txt
Flask==3.0.0
llama-index==0.9.8
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
python-multipart==0.0.6
